+++ receiver/jmxreceiver/receiver_test.go
@@ -35,7 +35,6 @@ func TestReceiver(t *testing.T) {
 		OTLPExporterConfig: otlpExporterConfig{
 			Endpoint: fmt.Sprintf("localhost:%d", testutil.GetAvailablePort(t)),
 		},
-		Properties: make(map[string]string),
 	}
 
 	receiver := newJMXMetricReceiver(params, config, consumertest.NewNop())
@@ -55,54 +54,10 @@ func TestBuildJMXMetricGathererConfig(t *testing.T) {
 		expectedError  string
 	}{
 		{
-			"uses target system",
-			Config{
-				Endpoint:           "service:jmx:rmi///jndi/rmi://myservice:12345/jmxrmi/",
-				TargetSystem:       "mytargetsystem",
-				GroovyScript:       "mygroovyscript",
-				CollectionInterval: 123 * time.Second,
-				OTLPExporterConfig: otlpExporterConfig{
-					Endpoint: "myotlpendpoint",
-					TimeoutSettings: exporterhelper.TimeoutSettings{
-						Timeout: 234 * time.Second,
-					},
-				},
-			},
-			`otel.jmx.service.url = service:jmx:rmi///jndi/rmi://myservice:12345/jmxrmi/
-otel.jmx.interval.milliseconds = 123000
-otel.jmx.target.system = mytargetsystem
-otel.metrics.exporter = otlp
-otel.exporter.otlp.endpoint = http://myotlpendpoint
-otel.exporter.otlp.timeout = 234000
-`, "",
-		},
-		{
-			"uses groovy script",
-			Config{
-				Endpoint:           "service:jmx:rmi///jndi/rmi://myservice:12345/jmxrmi/",
-				GroovyScript:       "mygroovyscript",
-				CollectionInterval: 123 * time.Second,
-				OTLPExporterConfig: otlpExporterConfig{
-					Endpoint: "http://myotlpendpoint",
-					TimeoutSettings: exporterhelper.TimeoutSettings{
-						Timeout: 234 * time.Second,
-					},
-				},
-			},
-			`otel.jmx.service.url = service:jmx:rmi///jndi/rmi://myservice:12345/jmxrmi/
-otel.jmx.interval.milliseconds = 123000
-otel.jmx.groovy.script = mygroovyscript
-otel.metrics.exporter = otlp
-otel.exporter.otlp.endpoint = http://myotlpendpoint
-otel.exporter.otlp.timeout = 234000
-`, "",
-		},
-		{
-			"uses endpoint as service url",
+			"handles all relevant input appropriately",
 			Config{
 				Endpoint:           "myhost:12345",
 				TargetSystem:       "mytargetsystem",
-				GroovyScript:       "mygroovyscript",
 				CollectionInterval: 123 * time.Second,
 				OTLPExporterConfig: otlpExporterConfig{
 					Endpoint: "https://myotlpendpoint",
@@ -114,22 +69,38 @@ otel.exporter.otlp.timeout = 234000
 						"three": "four",
 					},
 				},
+				// While these aren't realistic usernames/passwords, we want to test the
+				// multiline handling in place to reduce the attack surface of the
+				// interface to the JMX metrics gatherer
+				Username: "myuser\nname",
+				Password: `mypass 
+word`,
+				Realm:         "myrealm",
+				RemoteProfile: "myprofile",
+				ResourceAttributes: map[string]string{
+					"abc": "123",
+					"one": "two",
+				},
 			},
-			`otel.jmx.service.url = service:jmx:rmi:///jndi/rmi://myhost:12345/jmxrmi
-otel.jmx.interval.milliseconds = 123000
-otel.jmx.target.system = mytargetsystem
-otel.metrics.exporter = otlp
-otel.exporter.otlp.endpoint = https://myotlpendpoint
-otel.exporter.otlp.timeout = 234000
+			`otel.exporter.otlp.endpoint = https://myotlpendpoint
 otel.exporter.otlp.headers = one=two,three=four
-`, "",
+otel.exporter.otlp.timeout = 234000
+otel.jmx.interval.milliseconds = 123000
+otel.jmx.password = mypass \nword
+otel.jmx.realm = myrealm
+otel.jmx.remote.profile = myprofile
+otel.jmx.service.url = service:jmx:rmi:///jndi/rmi://myhost:12345/jmxrmi
+otel.jmx.target.system = mytargetsystem
+otel.jmx.username = myuser\nname
+otel.metrics.exporter = otlp
+otel.resource.attributes = abc=123,one=two`,
+			"",
 		},
 		{
 			"errors on portless endpoint",
 			Config{
 				Endpoint:           "myhostwithoutport",
 				TargetSystem:       "mytargetsystem",
-				GroovyScript:       "mygroovyscript",
 				CollectionInterval: 123 * time.Second,
 				OTLPExporterConfig: otlpExporterConfig{
 					Endpoint: "myotlpendpoint",
@@ -145,7 +116,6 @@ otel.exporter.otlp.headers = one=two,three=four
 			Config{
 				Endpoint:           "myhost:withoutvalidport",
 				TargetSystem:       "mytargetsystem",
-				GroovyScript:       "mygroovyscript",
 				CollectionInterval: 123 * time.Second,
 				OTLPExporterConfig: otlpExporterConfig{
 					Endpoint: "myotlpendpoint",
@@ -161,7 +131,6 @@ otel.exporter.otlp.headers = one=two,three=four
 			Config{
 				Endpoint:           ":::",
 				TargetSystem:       "mytargetsystem",
-				GroovyScript:       "mygroovyscript",
 				CollectionInterval: 123 * time.Second,
 				OTLPExporterConfig: otlpExporterConfig{
 					Endpoint: "myotlpendpoint",