+++ receiver/zookeeperreceiver/internal/metadata/generated_metrics_v2.go
@@ -5,6 +5,7 @@ package metadata
 import (
 	"time"
 
+	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 )
@@ -899,10 +900,11 @@ func newMetricZookeeperZnodeCount(settings MetricSettings) metricZookeeperZnodeC
 // MetricsBuilder provides an interface for scrapers to report metrics while taking care of all the transformations
 // required to produce metric representation defined in metadata and user settings.
 type MetricsBuilder struct {
-	startTime                                  pcommon.Timestamp // start time that will be applied to all recorded data points.
-	metricsCapacity                            int               // maximum observed number of metrics per resource.
-	resourceCapacity                           int               // maximum observed number of resource attributes.
-	metricsBuffer                              pmetric.Metrics   // accumulates metrics data before emitting.
+	startTime                                  pcommon.Timestamp   // start time that will be applied to all recorded data points.
+	metricsCapacity                            int                 // maximum observed number of metrics per resource.
+	resourceCapacity                           int                 // maximum observed number of resource attributes.
+	metricsBuffer                              pmetric.Metrics     // accumulates metrics data before emitting.
+	buildInfo                                  component.BuildInfo // contains version information
 	metricZookeeperConnectionActive            metricZookeeperConnectionActive
 	metricZookeeperDataTreeEphemeralNodeCount  metricZookeeperDataTreeEphemeralNodeCount
 	metricZookeeperDataTreeSize                metricZookeeperDataTreeSize
@@ -930,10 +932,11 @@ func WithStartTime(startTime pcommon.Timestamp) metricBuilderOption {
 	}
 }
 
-func NewMetricsBuilder(settings MetricsSettings, options ...metricBuilderOption) *MetricsBuilder {
+func NewMetricsBuilder(settings MetricsSettings, buildInfo component.BuildInfo, options ...metricBuilderOption) *MetricsBuilder {
 	mb := &MetricsBuilder{
 		startTime:                       pcommon.NewTimestampFromTime(time.Now()),
 		metricsBuffer:                   pmetric.NewMetrics(),
+		buildInfo:                       buildInfo,
 		metricZookeeperConnectionActive: newMetricZookeeperConnectionActive(settings.ZookeeperConnectionActive),
 		metricZookeeperDataTreeEphemeralNodeCount:  newMetricZookeeperDataTreeEphemeralNodeCount(settings.ZookeeperDataTreeEphemeralNodeCount),
 		metricZookeeperDataTreeSize:                newMetricZookeeperDataTreeSize(settings.ZookeeperDataTreeSize),
@@ -1013,6 +1016,7 @@ func (mb *MetricsBuilder) EmitForResource(rmo ...ResourceMetricsOption) {
 	rm.Resource().Attributes().EnsureCapacity(mb.resourceCapacity)
 	ils := rm.ScopeMetrics().AppendEmpty()
 	ils.Scope().SetName("otelcol/zookeeperreceiver")
+	ils.Scope().SetVersion(mb.buildInfo.Version)
 	ils.Metrics().EnsureCapacity(mb.metricsCapacity)
 	mb.metricZookeeperConnectionActive.emit(ils.Metrics())
 	mb.metricZookeeperDataTreeEphemeralNodeCount.emit(ils.Metrics())