+++ receiver/zookeeperreceiver/scraper_test.go
@@ -23,10 +23,12 @@ import (
 	"io/ioutil"
 	"net"
 	"path/filepath"
+	"runtime"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
@@ -46,6 +48,8 @@ type logMsg struct {
 func TestZookeeperMetricsScraperScrape(t *testing.T) {
 	tests := []struct {
 		name                         string
+		skip                         bool
+		skipComment                  string
 		expectedMetricsFilename      string
 		expectedResourceAttributes   map[string]string
 		metricsSettings              func() metadata.MetricsSettings
@@ -60,6 +64,8 @@ func TestZookeeperMetricsScraperScrape(t *testing.T) {
 	}{
 		{
 			name:                         "Test correctness with v3.4.14",
+			skip:                         runtime.GOOS == "windows",
+			skipComment:                  "skipping test on windows, see https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/10171",
 			mockedZKOutputSourceFilename: "mntr-3.4.14",
 			expectedMetricsFilename:      "correctness-v3.4.14",
 			expectedResourceAttributes: map[string]string{
@@ -208,6 +214,11 @@ func TestZookeeperMetricsScraperScrape(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			if tt.skip {
+				t.Log(tt.skipComment)
+				return
+			}
+
 			localAddr := testutil.GetAvailableLocalAddress(t)
 			if !tt.mockZKConnectionErr {
 				ms := mockedServer{ready: make(chan bool, 1)}
@@ -222,7 +233,9 @@ func TestZookeeperMetricsScraperScrape(t *testing.T) {
 			}
 
 			core, observedLogs := observer.New(zap.DebugLevel)
-			z, err := newZookeeperMetricsScraper(zap.New(core), cfg)
+			settings := componenttest.NewNopReceiverCreateSettings()
+			settings.Logger = zap.New(core)
+			z, err := newZookeeperMetricsScraper(settings, cfg)
 			require.NoError(t, err)
 			require.Equal(t, "zookeeper", z.Name())
 
@@ -270,7 +283,7 @@ func TestZookeeperMetricsScraperScrape(t *testing.T) {
 
 func TestZookeeperShutdownBeforeScrape(t *testing.T) {
 	cfg := createDefaultConfig().(*Config)
-	z, err := newZookeeperMetricsScraper(zap.NewNop(), cfg)
+	z, err := newZookeeperMetricsScraper(componenttest.NewNopReceiverCreateSettings(), cfg)
 	require.NoError(t, err)
 	require.NoError(t, z.shutdown(context.Background()))
 }